Clean up version-specific caches
The VersionSpecificCacheAndWrapperDistributionCleanupService introduced
by this commit scans `$GRADLE_USER_HOME/caches` for version-specific
cache directories. It then checks when each directory was last used
and deletes those that have not been used for 30 days. In addition,
it deletes the corresponding wrapper distributions in
`$GRADLE_USER_HOME/wrapper/dists`. Versions greater than or equal to the
current version are ignored because they might be using a different,
potentially conflicting, cleanup strategy.